rename `mod` add fill styles
