+++ app/src/data_type/mod.rs
@@ -1,9 +1,9 @@
 pub mod button_type;
+pub mod color;
 pub mod fill;
 pub mod shape;
-pub mod style;
 
 pub use button_type::ButtonType;
+pub use color::Color;
 pub use fill::Fill;
 pub use shape::Shape;
-pub use style::Color;
